[IOPAE-311] kebab-case renaming (#47)
* renaming

* add file in src to avoid build error

* rename ServiceReviewProxy in jiraProxy

* fix jira-client naming
