+++ apps/io-services-cms-webapp/src/_to-solve.ts
@@ -1,0 +1,9 @@
+/**
+ * WARNING
+ * ************************************************************************************
+ * The line of code below avoids an unexplained error that occurs in the project build.
+ * This file will be deleted as soon as we have a structured solution to the problem.
+ */
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import * as TE from "fp-ts/lib/TaskEither";
+// ************************************************************************************
