+++ apps/io-services-cms-webapp/src/utils/__tests__/jira-proxy.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it, vitest } from "vitest";
 import * as config from "../../config";
-import { Delegate, ServiceReviewProxy } from "../service_review_proxy";
-import { Service } from "@io-services-cms/models/service-lifecycle/types";
+import { Delegate, jiraProxy } from "../jira-proxy";
+import { ServiceLifecycle } from "@io-services-cms/models";
 import { NonEmptyString } from "@pagopa/ts-commons/lib/strings";
 import * as E from "fp-ts/lib/Either";
 import * as O from "fp-ts/lib/Option";
@@ -9,8 +9,8 @@ import {
   JiraAPIClient,
   JiraIssue,
   SearchJiraIssuesResponse,
-  jiraAPIClient,
-} from "../../jira_client";
+  jiraClient,
+} from "../../lib/clients/jira-client";
 
 const JIRA_CONFIG = {
   JIRA_NAMESPACE_URL: "anUrl",
@@ -55,7 +55,7 @@ const aService = {
     },
     require_secure_channel: false,
   },
-} as unknown as Service;
+} as unknown as ServiceLifecycle.definitions.Service;
 
 const aDelegate = {
   name: "Mario Rossi",
@@ -104,8 +104,8 @@ describe("Service Review Proxy", () => {
     );
     const mockFetch = getMockFetchWithStatus(201);
 
-    const aJiraClient: jiraAPIClient = JiraAPIClient(JIRA_CONFIG, mockFetch);
-    const proxy = ServiceReviewProxy(aJiraClient);
+    const aJiraClient: JiraAPIClient = jiraClient(JIRA_CONFIG, mockFetch);
+    const proxy = jiraProxy(aJiraClient);
     const serviceReviewTicket = await proxy.createJiraIssue(
       aService,
       aDelegate
@@ -129,8 +129,8 @@ describe("Service Review Proxy", () => {
     );
     const mockFetch = getMockFetchWithStatus(200);
 
-    const aJiraClient: jiraAPIClient = JiraAPIClient(JIRA_CONFIG, mockFetch);
-    const proxy = ServiceReviewProxy(aJiraClient);
+    const aJiraClient: JiraAPIClient = jiraClient(JIRA_CONFIG, mockFetch);
+    const proxy = jiraProxy(aJiraClient);
     const searchKeys = [
       "anIssueKey-1" as NonEmptyString,
       "anIssueKey-2" as NonEmptyString,
@@ -157,8 +157,8 @@ describe("Service Review Proxy", () => {
     );
     const mockFetch = getMockFetchWithStatus(200);
 
-    const aJiraClient: jiraAPIClient = JiraAPIClient(JIRA_CONFIG, mockFetch);
-    const proxy = ServiceReviewProxy(aJiraClient);
+    const aJiraClient: JiraAPIClient = jiraClient(JIRA_CONFIG, mockFetch);
+    const proxy = jiraProxy(aJiraClient);
     const serviceReview = await proxy.getJiraIssueByServiceId(
       "aServiceId" as NonEmptyString
     )();
@@ -189,8 +189,8 @@ describe("Service Review Proxy", () => {
     );
     const mockFetch = getMockFetchWithStatus(200);
 
-    const aJiraClient: jiraAPIClient = JiraAPIClient(JIRA_CONFIG, mockFetch);
-    const proxy = ServiceReviewProxy(aJiraClient);
+    const aJiraClient: JiraAPIClient = jiraClient(JIRA_CONFIG, mockFetch);
+    const proxy = jiraProxy(aJiraClient);
     const serviceReview = await proxy.getJiraIssueByServiceId(
       "aWrongServiceId" as NonEmptyString
     )();
@@ -215,8 +215,8 @@ describe("Service Review Proxy", () => {
     );
     const mockFetch = getMockFetchWithStatus(500);
 
-    const aJiraClient: jiraAPIClient = JiraAPIClient(JIRA_CONFIG, mockFetch);
-    const proxy = ServiceReviewProxy(aJiraClient);
+    const aJiraClient: JiraAPIClient = jiraClient(JIRA_CONFIG, mockFetch);
+    const proxy = jiraProxy(aJiraClient);
     const serviceReview = await proxy.getJiraIssueByServiceId(
       "aServiceId" as NonEmptyString
     )();
