+++ apps/io-services-cms-webapp/src/utils/jira-proxy.ts
@@ -5,11 +5,11 @@ import { pipe } from "fp-ts/lib/function";
 import { ServiceLifecycle } from "@io-services-cms/models";
 import {
   CreateJiraIssueResponse,
+  JiraAPIClient,
   JiraIssue,
   SearchJiraIssuesPayload,
   SearchJiraIssuesResponse,
-  jiraAPIClient,
-} from "../jira_client";
+} from "../lib/clients/jira-client";
 
 const formatOptionalStringValue = (value?: string) =>
   value || `{color:#FF991F}*[ DATO MANCANTE ]*{color}`;
@@ -17,7 +17,7 @@ const formatOptionalStringValue = (value?: string) =>
 const formatIssueTitle = (serviceId: NonEmptyString) =>
   `Review #${serviceId}` as NonEmptyString;
 
-export type ServiceReviewProxy = {
+export type JiraProxy = {
   readonly createJiraIssue: (
     service: ServiceLifecycle.definitions.Service,
     delegate: Delegate
@@ -37,9 +37,7 @@ export type Delegate = {
   permissions: Array<string | undefined>;
 };
 
-export const ServiceReviewProxy = (
-  jiraClient: jiraAPIClient
-): ServiceReviewProxy => {
+export const jiraProxy = (jiraClient: JiraAPIClient): JiraProxy => {
   const buildIssueCustomFields = (
     service: ServiceLifecycle.definitions.Service,
     delegate: Delegate
